## 2.0-ALPHA5

+++ src/main/java/com/lognex/api/entities/RetailDrawerCashInEntity.java
@@ -21,7 +21,7 @@ public class RetailDrawerCashInEntity extends DocumentEntity {
     private MetaEntity documents;
     private String created;
     private Boolean applicable;
-    private Integer sum;
+    private Long sum;
     private Integer version;
     private String moment;
     private String accountId;
